--- a/pkg/mtls/crypto/README
+++ b/pkg/mtls/crypto/README
@@ -4,6 +4,7 @@ mtls/crypto/tls
 2. Add transfer_tls.go that support smooth upgrade on TLS
 3. modify import paths
 4. Call TransferSetTLSInfo() to set transfer info when TLS handshake is successful
+5. reuse tls Conn's outBuf to reduce memory in large number of idle connection
 
 add supports for tls1.3-gm
 see draft:https://datatracker.ietf.org/doc/draft-yang-tls-tls13-sm-suites/