@cch123 2020-10-17
```
~/g/s/m/m/p/m/c/tls ❯❯❯ go test  -bench=ConnOutbufWithPool -benchmem -count=10 | tee old.txt
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/mtls/crypto/tls
BenchmarkConnOutbufWithPool-8   	 1443111	       770 ns/op	     149 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1322788	       766 ns/op	     162 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1399046	       773 ns/op	     153 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1420789	       755 ns/op	     151 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1453182	       793 ns/op	     148 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1514980	       775 ns/op	     141 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1469499	       752 ns/op	     146 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1471977	       798 ns/op	     146 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1461138	       770 ns/op	     147 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 1472355	       768 ns/op	     146 B/op	       0 allocs/op
PASS
ok  	mosn.io/mosn/pkg/mtls/crypto/tls	23.163s
```

```
~/g/s/m/m/p/m/c/tls ❯❯❯ go test -bench=ConnOutbufWithPool -benchmem -count=10 | tee new.txt
goos: darwin
goarch: amd64
pkg: mosn.io/mosn/pkg/mtls/crypto/tls
BenchmarkConnOutbufWithPool-8   	 5121081	       216 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5493721	       209 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5279673	       209 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5575989	       211 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5201364	       206 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5393505	       214 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5406199	       211 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5049882	       209 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5485282	       199 ns/op	       2 B/op	       0 allocs/op
BenchmarkConnOutbufWithPool-8   	 5613744	       209 ns/op	       2 B/op	       0 allocs/op
PASS
ok  	mosn.io/mosn/pkg/mtls/crypto/tls	21.061s
```

```
~/g/s/m/m/p/m/c/tls ❯❯❯ benchstat old.txt new.txt
name                  old time/op    new time/op    delta
ConnOutbufWithPool-8     769ns ± 3%     210ns ± 2%  -72.73%  (p=0.000 n=9+8)

name                  old alloc/op   new alloc/op   delta
ConnOutbufWithPool-8      147B ± 4%        2B ± 0%  -98.64%  (p=0.000 n=9+10)

name                  old allocs/op  new allocs/op  delta
ConnOutbufWithPool-8      0.00           0.00          ~     (all equal)
~/g/s/m/m/p/m/c/tls ❯❯❯
```