@cch123 2020-10-23
Detailed Benchmark:

## Max QPS test

GOMAXPROCS=16. / wrk -c 1024 -d 10s -t64 https://localhost:4443

Original TLS:

|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 249252.58 | 247981.13 | 252072.94|  246838.65|  248069.50|  250279.61|  248893.85 |  248653.03 | 

Use sync.Pool:

|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 233128.73 | 236434.06 | 241020.74 | 237904.41 | 237683.17|  240238.59 | 241162.77 |  238880.61|

GOMAXPROCS=8. / wrk -c 1024 -d 10s -t64 https://localhost:4443

Original TLS

|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 137442.72| 135192.82 |  137283.12| 138932.01| 136880.19 |  137972.56|  139443.37 |  140283.15|


Use sync.Pool:
|seq|1|2|3|4|5|6|7|8
|--|--|--|--|--|--|--|--|--|
| QPS | 134797.43 |132957.91| 132796.81| 134416.20| 136944.84| 135315.26| 134220.55| 133366.97|



## Fixed Connections, Fixed QPS, Max memory test

```
$sudo ./nbclient 50000 3000 https://ip:4443
```

```
$sudo netstat -antp |grep 4443  |wc -l
50001
```

Original TLS:
```
   PID USER      PR  NI    VIRT    RES    SHR S  %CPU %MEM     TIME+ COMMAND
 39447 root      20   0 3833636   3.0g   3864 S  27.9  1.2   4:09.11 server_no_pool
  8018 root      20   0 3686144  27900   4476 S   2.0  0.0   4547:52 y_watch
```

Use sync.Pool:
```
   PID USER      PR  NI    VIRT    RES    SHR S  %CPU %MEM     TIME+ COMMAND
 42347 root      20   0 2598440   1.8g   3900 S  29.8  0.7   4:39.22 server_with_pool
  8018 root      20   0 3686144  27900   4476 S   2.0  0.0   4547:56 y_watchdog
```

## Our real scene test

We are using Go to develop our mesh project, [mosn](github.com/mosn/mosn), and we are running mosn online in 8 core environment, 

So the request path is wrk --(https)---> mosn -> normal http server

the following are the max QPS before and after outBuf change:

Original TLS:

|seq|1|2|3|4|5|
|--|--|--|--|--|--|
| QPS | 53408.99 | 52830.87 | 53602.04 | 53519.22 | 53251.51 |

Use sync.Pool:

|seq|1|2|3|4|5|
|--|--|--|--|--|--|
| QPS | 53468.03 | 54663.48 |  53911.13 | 53410.65 | 53669.69 |

As long as there are some real network interactions and business logic, the max QPS has little performance difference.

## server and client

https server:
```go
package main

import (
	"fmt"
	"net/http"
	_ "net/http/pprof"
)

var content = make([]byte, 16000)

func sayhello(wr http.ResponseWriter, r *http.Request) {
	wr.Header()["Content-Length"] = []string{fmt.Sprint(len(content))}
	wr.Header()["Content-Type"] = []string{"application/json"}
	wr.Write(content)
}

func main() {
	go func() {
		http.ListenAndServe(":3333", nil)
	}()
	http.HandleFunc("/", sayhello)

	err := http.ListenAndServeTLS(":4443", "server.crt", "server.key", nil)
	if err != nil {
		fmt.Println(err)
	}
}
```

fixed qps client:
```go
package main

import (
	"crypto/tls"
	"fmt"
	"io/ioutil"
	"net/http"
	"os"
	"strconv"
	"sync"

	"go.uber.org/ratelimit"
)

func main() {
	url := os.Args[3]
	connNum, err := strconv.ParseInt(os.Args[1], 10, 64)
	if err != nil {
		fmt.Println(err)
		return
	}

	qps, err := strconv.ParseInt(os.Args[2], 10, 64)
	if err != nil {
		fmt.Println(err)
		return
	}

	bucket := ratelimit.New(int(qps))

	var l sync.Mutex
	connList := make([]*http.Client, connNum)

	for i := 0; ; i++ {
		bucket.Take()
		i := i
		go func() {
			l.Lock()
			if connList[i%len(connList)] == nil {
				connList[i%len(connList)] = &http.Client{
					Transport: &http.Transport{
						TLSClientConfig:     &tls.Config{InsecureSkipVerify: true},
						IdleConnTimeout:     0,
						MaxIdleConns:        1,
						MaxIdleConnsPerHost: 1,
					},
				}
			}
			conn := connList[i%len(connList)]
			l.Unlock()
			if resp, e := conn.Get(url); e != nil {
				fmt.Println(e)
			} else {
				defer resp.Body.Close()
				ioutil.ReadAll(resp.Body)
			}
		}()
	}

}

```