@taoyuanyuan 2020-10-26
https://github.com/golang/go/pull/42036